re-organizes subject list into clusters
