+++ tests/dnaTests.js
@@ -1,15 +1,6 @@
 var path = require('path');
 var diagnosticsTestSuite = require('./genosets/meta-data');
-
-//TEST SUBJECTS
-var rootPath = path.join(__dirname, '../');
-var dna = require(path.join(rootPath,'data/json/23andme-male.json'));
-var dna2 = require(path.join(rootPath,'data/json/ancestry-female.json'));
-var dna3 = require(path.join(rootPath, 'data/json/family-tree-male.json'));
-var dna4 = require(path.join(rootPath, 'data/json/user-upload.json'));
-
-//SUBJECT LIST
-var testSubjects = [dna, dna2, dna3, dna4];
+var testSubjects = require('./subjects/cluster');
 
 //DIAGNOSTIC TESTS RUNNER
 function runDNAJSON (testSuite, subject){
@@ -27,10 +18,7 @@ function runDNAJSON (testSuite, subject){
 }
 
 
-//INDIVIDUAL TESTING SUITE TYPE RUNNERS
-
-//ABSTRACT THESE TESTS so they all run the same with just small differences
-//ex:
+//INDIVIDUAL TESTING SUITE RUNNER
 function testRunner (tests, subject, testingType){
 	return tests.map(function(diagnosticTest){
 		var testObject = {
@@ -52,10 +40,6 @@ function aggregateTestData (results, factor) {
 	var localAccumulator = 0;
 	results.forEach(function (result){
 		if(result.booleanVal){
-
-			//Calculate health score totals
-			//localAccumulator is the result of score combos for a specific test suite (highrisk, lowrisk, prevention, immunity etc);
-
 			//take the result.healthFactor (either +1 || -1) * the result[factor] either riskMultiplier or magnitue for prevention tests
 			//just letting us sub in the factor depending on the testing suite
 			localAccumulator += (Number(result.healthFactor) * result[factor]);
